+++ Assets/Scripts/Damage/ElementalWeakness.cs
@@ -7,11 +7,11 @@ namespace Damage
         public Element weakness;
         public float multiplier = 2;
         
-        protected override int Modifier(int damage, Element element)
+        protected override float Modifier(float damage, Element element)
         {
             if (element == weakness)
             {
-                return (int) Math.Round(damage * multiplier);
+                return (float) Math.Round(damage * multiplier);
             }
             return damage;
         }
